% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crs_gda2020}
\alias{crs_gda2020}
\title{Coordinate reference system for Australia}
\usage{
crs_gda2020()
}
\value{
A simple features CRS
}
\description{
GDA2020 is the official CRS used by the
\href{https://www.icsm.gov.au/gda2020}{Commonwealth} and
\href{https://www.spatial.nsw.gov.au/surveying/geodesy/gda2020}{NSW}.
Geospatial data in this package uses GDA2020.
}
