% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalise_state_names}
\alias{normalise_state_names}
\title{Normalise state names from abbreviations}
\usage{
normalise_state_names(names)
}
\arguments{
\item{names}{Character vector of state names.}
}
\value{
Vector of the same size as the input, but with the normalised state names.
}
\description{
Expand abbreviations like \code{"NSW"} to \code{"New South Wales"}, and
normalise to title capitalisation. Entries that don't match any state name or
abbreviation are left untouched.
}
\examples{
normalise_state_names(c("nsw", "VIC", "overseas", "Queensland"))
}
