% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phn}
\alias{phn}
\title{Primary Health Network boundaries of New South Wales}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10 rows and 9 columns.
}
\source{
National Recovery and Resilience Agency, "PHN Boundaries used by the NBRA",
\url{https://data.gov.au/data/dataset/phn-boundaries-used-by-the-nbra}, accessed 25 May 2023.

The original dataset is published under the \href{https://creativecommons.org/licenses/by/2.5/au/}{Creative Commons Attribution 2.5 Australia} licence,
© Commonwealth of Australia 2021.
}
\usage{
phn
}
\description{
The geometries have been simplified with a tolerance of 500 m to reduce the
level of detail.
}
\keyword{datasets}
