% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{suburbs}
\alias{suburbs}
\title{Suburbs of New South Wales.}
\format{
A data frame with 2 columns:
\describe{
\item{suburbname}{The name of the suburb, in upper case}
\item{postcode}{The main postcode of the suburb, as a character}
}
}
\source{
Spatial Services, Department of Customer Service NSW.
"NSW Administrative Boundaries Theme - Suburb."
\url{https://portal.spatial.nsw.gov.au/portal/home/item.html?id=38bdaa10b7cc41a3a19be6eca91f5368}, accessed 21 September 2022.

The original dataset is published under the \href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons Attribution 4.0 International} licence,
© State of New South Wales (\href{https://www.spatial.nsw.gov.au/copyright}{Spatial Services}, a business unit of the Department of Customer Service NSW).
}
\usage{
suburbs
}
\description{
A dataset containing the names of suburbs in NSW and their postcodes.
These fields are extracted as-is from the source dataset published by DCS Spatial Services, NSW Government.
}
\keyword{datasets}
